requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.54.0
django-crispy-forms==2.0
gunicorn==20.1.0
heroicons==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==9.5.0
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.1
toml==0.10.2
Unipath==1.1
urllib3==2.0.3
whitenoise==5.3.0
